requirements.txt
pdfplumber
pdf2image
pytesseract
python-docx
ebooklib
beautifulsoup4
pillow
requests